+++ tests/testthat/test-expression.R
@@ -43,14 +43,25 @@ with_mock_HTTP({
     })
 
     test_that("R logical & CrunchLogicalExpr", {
-        e <- c(TRUE, FALSE, TRUE) & ds$gender == "Female"
-        expect_is(e, "CrunchLogicalExpr")
-        e <- c(TRUE, FALSE, TRUE) | ds$gender == "Female"
-        expect_is(e, "CrunchLogicalExpr")
-        e <- ds$gender == "Female" & c(TRUE, FALSE, TRUE)
-        expect_is(e, "CrunchLogicalExpr")
-        e <- ds$gender == "Female" | c(TRUE, FALSE, TRUE)
-        expect_is(e, "CrunchLogicalExpr")
+        expect_is(c(TRUE, FALSE, TRUE) & ds$gender == "Female",
+            "CrunchLogicalExpr")
+        expect_is(c(TRUE, FALSE, TRUE) | ds$gender == "Female",
+            "CrunchLogicalExpr")
+        expect_is(ds$gender == "Female" & c(TRUE, FALSE, TRUE),
+            "CrunchLogicalExpr")
+        expect_is(ds$gender == "Female" | c(TRUE, FALSE, TRUE),
+            "CrunchLogicalExpr")
+    })
+
+    test_that("Datetime operations: logical", {
+        expect_output(ds$starttime == "2015-01-01",
+            'Crunch logical expression: starttime == "2015-01-01"')
+        expect_output(ds$starttime > "2015-01-01",
+            'Crunch logical expression: starttime > "2015-01-01"')
+        expect_output(ds$starttime == as.Date("2015-01-01"),
+            'Crunch logical expression: starttime == "2015-01-01"')
+        expect_output(ds$starttime > as.Date("2015-01-01"),
+            'Crunch logical expression: starttime > "2015-01-01"')
     })
 
     test_that("Referencing category names that don't exist errors", {
@@ -77,8 +88,6 @@ with_mock_HTTP({
         expect_output(ds$gender %in% "Male" & !is.na(ds$birthyr),
             'gender == "Male" & !is.na(birthyr)',
             fixed=TRUE)
-        skip("TODO: implement datetime ops first")
-        print(ds$starttime > "2015-04-01")
     })
     test_that("Show method for expresssions", {
         skip("TODO: something intelligent with parentheses and order of operations")